backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3

sentence-transformers==4.1.0
faiss-cpu==1.12.0
numpy==2.3.0
PyMuPDF==1.26.1
openai==1.93.0
python-multipart==0.0.20
## The following requirements were added by pip freeze:
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
filelock==3.25.2
fsspec==2026.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
packaging==26.0
pillow==12.2.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3